flowfixPyAgent/pyproject.toml
[project]
name = "flowfix"
version = "0.1.0"
description = "FlowFix AI Agent + RAG System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "psycopg2-binary>=2.9.9",
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.0",
    "pika>=1.3.2",
    "pgvector>=0.2.0",
    "psycopg>=3.1.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "structlog>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
